+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/ui/actions/scan/InvalidateCachesAction.kt
@@ -4,8 +4,13 @@ import com.github.warningimhack3r.npmupdatedependencies.backend.engine.NUDState
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.diagnostic.logger
 
 class InvalidateCachesAction : AnAction() {
+    companion object {
+        private val log = logger<InvalidateCachesAction>()
+    }
+
     override fun getActionUpdateThread(): ActionUpdateThread {
         return ActionUpdateThread.BGT
     }
@@ -18,8 +23,11 @@ class InvalidateCachesAction : AnAction() {
     }
 
     override fun actionPerformed(e: AnActionEvent) {
-        val state = e.project?.let { NUDState.getInstance(it) } ?: return
-        state.availableUpdates.clear()
-        state.deprecations.clear()
+        log.debug("Cache invalidation requested")
+        val state = e.project?.let { NUDState.getInstance(it) } ?: return.also {
+            log.warn("No project found")
+        }
+        state.invalidateCaches()
+        log.debug("Cache invalidated")
     }
 }